requirements.txt
alembic==1.7.5
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Limiter==2.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.3.1
Mako==1.1.6
MarkupSafe==2.0.1
pydantic==1.9.0
PyJWT==2.3.0
shortuuid==1.0.8
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.0.1
Werkzeug==2.0.2
